PosixSourceAccessor: always require a root
With the introduction of CanonPath, we don't have roots on paths we
want to access. For example, /nix/store/abc doesn't say whether we
need to use C: or D: on Windows.

This commit keeps getFSSourceAccessor() which does NOT have a root
specified, so it pulls one from the current working directory. This is
silly, since nix.exe could be on Z: but the Nix store could be on C:
which means we'll build wrong paths. More commits will have to come
which removes getFSSourceAccessor() altogether.

This shouldn't change anything on Unix systems, since the root should
always resolve to / for every possible path.