lockFile: alternative implementation
The previous implementation looked a bit suspicious to me, not sure if
it was doing the right thing at all. It tries to set locks on two
bytes of a file, but I think that fails when trying to lock an empty
file.

This alternative implementation is a lot more simple and should
reflect the "flock" calls used under the Unix implementation.

But we need some automated tests to make sure this alternative
implementation works at all.