rapidcheck: switch to combinators instead of dereference operator
For some reason on Windows, the arbitrary instances are being executed
with an improper Rapidcheck handler on the stack. When trying to
dereference on Windows, we get:

    operator* is not allowed in this context

https://github.com/emil-e/rapidcheck/blob/ff6af6fc683159deb51c543b065eba14dfcf329b/src/gen/detail/GenerationHandler.cpp#L16C31-L16C71

I haven't been able to figure out why, but I figured that just not
using the operator is possible, so let's go with that.